@@ -3,8 +3,8 @@
   "description": "React",
   "version": "1.0.0",
   "scripts": {
-    "build": "liferay-npm-bundler-improved",
-    "deploy": "liferay-npm-bundler-improved -d /opt/liferayPlaygroundDXP73/deploy"
+    "dev:build": "liferay-npm-bundler-improved",
+    "dev:deploy": "liferay-npm-bundler-improved -d /opt/liferay/deploy"
   },
   "portlet": {
     "com.liferay.portlet.display-category": "category.sample",
@@ -13,22 +13,21 @@
     "javax.portlet.name": "react",
     "javax.portlet.display-name": "React"
   },
-  "main": "index.js",
+  "main": "index",
   "dependencies": {
     "react": "17.0.2",
     "react-dom": "17.0.2",
-    "tslib": "^2.2.0"
+    "tslib": "^2.3.1"
   },
   "devDependencies": {
     "@rollup/plugin-commonjs": "^19.0.0",
     "@rollup/plugin-node-resolve": "^13.0.0",
-    "@rollup/plugin-replace": "^2.4.2",
     "@rollup/plugin-typescript": "^8.2.1",
     "@types/react": "^17.0.11",
     "@types/react-dom": "^17.0.7",
-    "liferay-npm-bundler-improved": "^1.0.0-beta.6",
-    "rollup-plugin-peer-deps-external": "^2.2.4",
-    "rollup-plugin-terser": "^7.0.2",
-    "typescript": "^4.3.2"
+    "esbuild": "^0.12.27",
+    "liferay-npm-bundler-improved": "^1.0.0-beta.7",
+    "rollup-plugin-esbuild": "^4.5.0",
+    "typescript": "^4.4.2"
   }
 }
