@@ -1,58 +1,56 @@
 #!/usr/bin/env node
-const yargs = require('yargs')
-
-import Bundler from './classes/Bundler'
+import TimeHandler from './handlers/TimeHandler'
+import yargs from 'yargs'
 import Log from './classes/Log'
+import Bundler from './classes/Bundler'
 
-async function initialization () {
-  const start: Date = new Date()
+const arg: any = yargs(process.argv.slice(2)).options({
+  deploy: {
+    alias: 'd',
+    type: 'string'
+  },
+  keep: {
+    alias: 'k',
+    type: 'boolean'
+  }
+}).argv;
 
-  Log.write(Log.chalk.blue('liferay-npm-bundler-improved - 1.0.0-beta.7'))
+(async () => {
+  const timer = new TimeHandler()
+  Log.titleBadge(false, 'LIFERAY-NPM-BUNDLER-IMPROVED')
+  Log.blackBadge(false, '1.0.0-beta.7')
 
   const bundler: Bundler = new Bundler()
 
-  const arg: any = yargs(process.argv.slice(2)).options({
-    deploy: {
-      alias: 'd',
-      type: 'string'
-    },
-    keep: {
-      alias: 'k',
-      type: 'boolean'
-    }
-  }).argv
-
   try {
-    Log.write('\npreparation')
-    await bundler.loadRollupConfiguration()
-    await bundler.createDistDirectory()
+    Log.mainBadge(true, 'prepare')
+    await bundler.prepare()
 
-    Log.write('\nbundle code with rollup')
+    Log.mainBadge(true, 'bundle')
     await bundler.bundle()
 
-    Log.write('\nwrap liferay code into the Liferay.Loader')
-    await bundler.wrap()
+    Log.mainBadge(true, 'process')
+    await bundler.process()
 
-    Log.write('\ncheck additional portlet features')
-    await bundler.features()
-
-    Log.write('\ncreate and save jar')
+    Log.mainBadge(true, 'create')
     await bundler.create()
 
     if (!arg.keep) {
-      Log.write('\ncleanup')
+      Log.mainBadge(true, 'cleanup')
       await bundler.cleanup()
     }
 
     if (arg.deploy || arg.deploy === '') {
-      Log.write('\nadditional: deploying')
+      Log.mainBadge(true, 'deploy')
       await bundler.deploy(arg.deploy)
     }
 
-    Log.write(Log.chalk.bgGreen(`\nbundler finished successfully`), Log.chalk.bgBlack(`took ${(new Date().getTime() - start.getTime()) / 1000 }s`))
+    Log.successBadge(true, 'success')
+    Log.success(timer, 'bundler finished successfully')
+    process.exit(0)
   } catch (exception) {
-    Log.write(Log.chalk.bgRed(`\nbundler finished with an error`), Log.chalk.bgBlack(`took ${(new Date().getTime() - start.getTime()) / 1000 }s`))
+    Log.errorBadge(true, 'error')
+    Log.error(timer, 'bundler finished with an error')
+    process.exit(1)
   }
-}
-
-initialization()
+})()
