@@ -1,32 +1,19 @@
 import commonjs from '@rollup/plugin-commonjs'
 import resolve from '@rollup/plugin-node-resolve'
-import replace from '@rollup/plugin-replace'
-import typescript from '@rollup/plugin-typescript'
-import { terser } from 'rollup-plugin-terser'
-import peerDepsExternal from 'rollup-plugin-peer-deps-external'
+import esbuild from 'rollup-plugin-esbuild'
 
 export default {
   input: 'src/index.tsx',
   output: [
     {
       file: 'dist/index.js',
-      format: 'commonjs'
+      format: 'commonjs',
+      exports: 'default'
     },
   ],
   plugins: [
-    peerDepsExternal(),
     resolve(),
-    replace({
-      'process.env.NODE_ENV': JSON.stringify('production'),
-      preventAssignment: true
-    }),
-    commonjs({
-      sourceMap: false
-    }),
-    typescript({
-      sourceMap: false,
-      inlineSources: false
-    }),
-    terser()
+    commonjs(),
+    esbuild()
   ]
 }
