@@ -4,32 +4,28 @@ import LiferayParams from '../types/LiferayParams'
 declare const Liferay: any
 
 const AppComponent = (liferayParams: LiferayParams): any => {
-  const {
-    portletNamespace, contextPath, portletElementId, configuration,
-  } = liferayParams
-
   return (
     <div>
       <div>
         <span className="tag">
-          {Liferay.Language.get('portlet-namespace')} Deger
+          {Liferay.Language.get('portlet-namespace')}
           :
         </span>
-        <span className="value">{portletNamespace}</span>
+        <span className="value">{liferayParams.portletNamespace}</span>
       </div>
       <div>
         <span className="tag">
           {Liferay.Language.get('context-path')}
           :
         </span>
-        <span className="value">{contextPath}</span>
+        <span className="value">{liferayParams.contextPath}</span>
       </div>
       <div>
         <span className="tag">
           {Liferay.Language.get('portlet-element-id')}
           :
         </span>
-        <span className="value">{portletElementId}</span>
+        <span className="value">{liferayParams.portletElementId}</span>
       </div>
       <div>
         <span className="tag">
@@ -37,7 +33,7 @@ const AppComponent = (liferayParams: LiferayParams): any => {
           :
         </span>
         <span className="value pre">
-          {JSON.stringify(configuration, null, 2)}
+          {JSON.stringify(liferayParams.configuration, null, 2)}
         </span>
       </div>
     </div>
