@@ -1,4 +1,4 @@
-import Vue from 'vue/dist/vue.common'
+import Vue from 'vue'
 import App from './components/index.vue'
 
 export default function main ({ portletNamespace, contextPath, portletElementId, configuration }) {
